+++ frontend/src/components/record/PeriodSelectionBar/PeriodSelectionBar.tsx
@@ -2,67 +2,44 @@ import { useState } from 'react';
 import { css, styled } from 'styled-components';
 
 import Button from '@Components/common/Button/Button';
-import Typography from '@Components/common/Typography/Typography';
 
-import useCalendar from '@Hooks/common/useCalendar';
 import useOutsideClick from '@Hooks/common/useOutsideClick';
 
 import color from '@Styles/color';
 
 import { PERIOD } from '@Constants/record';
 
-import ArrowIcon from '@Assets/icons/ArrowIcon';
 import CalenderIcon from '@Assets/icons/CalenderIcon';
 
-import format from '@Utils/format';
-
-import CalendarDayOfWeeks from '../CalendarDayOfWeeks/CalendarDayOfWeeks';
+import PeriodSelectCalendar from '../PeriodSelectCalendar/PeriodSelectCalendar';
 import type { Period } from '../contexts/MemberRecordPeriodProvider';
 import { useMemberRecordPeriod } from '../contexts/MemberRecordPeriodProvider';
 
 const periodTypes: Period[] = ['week', 'oneMonth', 'threeMonth', 'entire'];
 
 const PeriodSelectionBar = () => {
-  const today = new Date();
+  const { period, startDate, endDate, hasSelectedCustomPeriod, handlePeriod } = useMemberRecordPeriod();
 
-  const {
-    period,
-    startDate,
-    endDate,
-    hasSelectedCustomPeriod,
-    isMiddleSelectedCustomDate,
-    handlePeriod,
-    handleCustomPeriod,
-    handleHoverDays,
-  } = useMemberRecordPeriod();
-
-  const { year, month, monthStorage, handleMonthShift } = useCalendar();
-  const [isOpenSelectPeriodDate, setIsOpenSelectPeriodDate] = useState(false);
+  const [isOpenPeriodSelectCalendar, setIsOpenPeriodSelectCalendar] = useState(false);
 
   const ref = useOutsideClick<HTMLDivElement>(() => {
-    setIsOpenSelectPeriodDate(false);
+    setIsOpenPeriodSelectCalendar(false);
   });
-
-  const getDayBackgroundColor = (date: string) => {
-    if (startDate === date || endDate === date) return color.blue[200];
-
-    if (isMiddleSelectedCustomDate(date)) return color.blue[100];
-
-    if (date === format.date(today)) return color.neutral[100];
-
-    return 'transparent';
-  };
 
   return (
     <Layout>
-      <SelectButtonContainer>
+      <SelectPeriodButtonContainer>
         {periodTypes.map((periodType) => (
-          <SelectButton $isSelected={period === periodType} onClick={() => handlePeriod(periodType)} key={periodType}>
+          <SelectPeriodButton
+            $isSelected={period === periodType}
+            onClick={() => handlePeriod(periodType)}
+            key={periodType}
+          >
             {PERIOD[periodType]}
-          </SelectButton>
+          </SelectPeriodButton>
         ))}
-      </SelectButtonContainer>
-      <UserSelectPeriodContainer>
+      </SelectPeriodButtonContainer>
+      <SelectCustomPeriodContainer>
         <SelectDateWrapper ref={ref}>
           <SelectedDate $hasSelectedCustomPeriod={hasSelectedCustomPeriod}>
             {hasSelectedCustomPeriod ? (
@@ -75,41 +52,15 @@ const PeriodSelectionBar = () => {
               '날짜를 선택해주세요.'
             )}
           </SelectedDate>
-          <SelectDateButton onClick={() => setIsOpenSelectPeriodDate(true)}>
+          <SelectDateButton onClick={() => setIsOpenPeriodSelectCalendar(true)}>
             <CalenderIcon color={color.black} />
           </SelectDateButton>
-          {isOpenSelectPeriodDate && (
-            <SelectPeriod>
-              <Month>
-                <Typography variant="p2">
-                  {year}년 {month}월
-                </Typography>
-                <div>
-                  <ArrowIcon direction="left" onClick={() => handleMonthShift('prev')} />
-                  <ArrowIcon direction="right" onClick={() => handleMonthShift('next')} />
-                </div>
-              </Month>
-              <CalendarDayOfWeeks position="center" />
-              <Days>
-                {monthStorage.map(({ day, fullDate, state, fullDateDot }) => (
-                  <Day
-                    key={fullDate}
-                    $isCurrentMonthDay={state === 'cur'}
-                    onClick={() => handleCustomPeriod(fullDateDot)}
-                    onMouseEnter={() => handleHoverDays(fullDateDot)}
-                    $backgroundColor={getDayBackgroundColor(fullDateDot)}
-                  >
-                    {day}
-                  </Day>
-                ))}
-              </Days>
-            </SelectPeriod>
-          )}
+          {isOpenPeriodSelectCalendar && <PeriodSelectCalendar />}
         </SelectDateWrapper>
         <Button variant="outlined" size="x-small" onClick={() => handlePeriod(null)}>
           조회
         </Button>
-      </UserSelectPeriodContainer>
+      </SelectCustomPeriodContainer>
     </Layout>
   );
 };
@@ -124,7 +75,7 @@ const Layout = styled.div`
   user-select: none;
 `;
 
-const SelectButtonContainer = styled.div`
+const SelectPeriodButtonContainer = styled.div`
   display: flex;
 
   border-radius: 8px;
@@ -143,11 +94,11 @@ const SelectButtonContainer = styled.div`
   }
 `;
 
-type SelectButtonProps = {
+type SelectPeriodButtonProps = {
   $isSelected: boolean;
 };
 
-const SelectButton = styled.button<SelectButtonProps>`
+const SelectPeriodButton = styled.button<SelectPeriodButtonProps>`
   padding: 10px 20px;
 
   border: 1px solid transparent;
@@ -158,7 +109,7 @@ const SelectButton = styled.button<SelectButtonProps>`
   `}
 `;
 
-const UserSelectPeriodContainer = styled.div`
+const SelectCustomPeriodContainer = styled.div`
   display: flex;
   gap: 10px;
   align-items: center;
@@ -228,81 +179,4 @@ const SelectDateButton = styled.div`
     width: 20px;
     height: 20px;
   }
-`;
-
-const SelectPeriod = styled.div`
-  position: absolute;
-  top: 50px;
-  right: 0;
-  left: 0;
-
-  background-color: ${color.white};
-
-  padding: 15px;
-  border: 1px solid ${color.neutral[100]};
-  border-radius: 4px;
-
-  box-shadow: rgba(149, 157, 165, 0.2) 0px 8px 24px;
-
-  z-index: 5;
-`;
-
-const Month = styled.div`
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-  gap: 10px;
-
-  padding: 0px 10px;
-  margin-bottom: 15px;
-
-  p {
-    font-weight: 500;
-  }
-
-  div {
-    display: flex;
-    gap: 20px;
-  }
-
-  svg {
-    cursor: pointer;
-  }
-`;
-
-const Days = styled.ul`
-  display: grid;
-  row-gap: 5px;
-  grid-template-columns: repeat(7, 1fr);
-  justify-content: center;
-
-  margin-top: 10px;
-`;
-
-type DayProps = {
-  $isCurrentMonthDay: boolean;
-  $backgroundColor: string;
-};
-
-const Day = styled.li<DayProps>`
-  position: relative;
-
-  display: flex;
-  justify-content: center;
-  align-items: center;
-
-  padding: 5px 10px;
-  text-align: center;
-
-  width: 42px;
-  height: 42px;
-
-  cursor: pointer;
-
-  transition: background-color 0.1s ease;
-
-  ${({ $isCurrentMonthDay, $backgroundColor }) => css`
-    opacity: ${$isCurrentMonthDay ? 1 : 0.4};
-    background-color: ${$backgroundColor};
-  `}
 `;
